+++ containers/Expected.h
@@ -46,7 +46,7 @@ namespace evo{
 
 			EVO_NODISCARD constexpr Expected() noexcept : expected_val(), has_expected_value(true) {};
 
-			EVO_NODISCARD constexpr ~Expected() noexcept {
+			constexpr ~Expected() noexcept {
 				if(this->has_value()){
 					std::destroy_at(&this->expected_val);
 				}else{